update for eld 0.1.6

+++ src/main/java/com/ericlam/mc/eldtester/TesterLifeCycle.java
@@ -1,8 +1,10 @@
 package com.ericlam.mc.eldtester;
 
 import com.ericlam.mc.eld.ELDLifeCycle;
+import com.ericlam.mc.eld.services.LoggingService;
 import com.ericlam.mc.eldtester.sql.User;
 import com.ericlam.mc.eldtester.sql.UserService;
+import org.bukkit.plugin.PluginLogger;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import javax.inject.Inject;
@@ -15,21 +17,44 @@ public class TesterLifeCycle implements ELDLifeCycle {
     @Inject
     private UserService userService;
 
+    @Inject
+    private LoggingService loggingService;
+
     @Override
     public void onEnable(JavaPlugin javaPlugin) {
+        var exist = userService.existByUsername("1234");
+        javaPlugin.getLogger().info("exist 1234: "+exist);
+        if (exist) {
+            userService.deleteById("1234");
+            javaPlugin.getLogger().info("after delete: ");
+            userService.findAll().forEach(u -> javaPlugin.getLogger().info(u.toString()));
+        }
         javaPlugin.getLogger().info("exist 1234: "+userService.existByUsername("1234"));
-        userService.deleteById("1234");
-        javaPlugin.getLogger().info("after delete: ");
-        userService.findAll().forEach(u -> javaPlugin.getLogger().info(u.toString()));
-        javaPlugin.getLogger().info("exist 1234: "+userService.existByUsername("1234"));
-        User user = new User();
-        user.username = "1234";
-        user.password = "1234";
-        user.lastName = "A";
-        user.firstName = "B";
-        userService.save(user);
-        javaPlugin.getLogger().info("after create: ");
-        userService.findAll().forEach(u -> javaPlugin.getLogger().info(u.toString()));
+
+
+
+        // save
+        var u = userService.findByUsername("1234").orElseGet(() -> {
+            var user = new User();
+            user.username = "1234";
+            user.password = "1234";
+            user.firstName = "Eric";
+            user.lastName = "Lam";
+            return user;
+        });
+
+        javaPlugin.getLogger().info("current: "+ u);
+
+        u.firstName = "Eric2";
+        u.lastName = "Lam2";
+
+        userService.save(u);
+
+        var logger = loggingService.getLogger(ELDTester.class);
+        logger.debug("this {0} is {1}", "logger", "debug logger");
+        logger.debug(new Exception("this is a debug exception"), "this is a debug {0}", "exception");
+        logger.info("this is a info message");
+        logger.warn("this is a warn message");
     }
 
     @Override
